backend/requirements.txt
fastapi
uvicorn
pydantic
spacy
PyPDF2
python-multipart
pymongo
python-dotenv
mongomock
email-validator